Improved params vs. fields types.
Parameters are now borrowed values from the caller. Including
`&str` for string and borrowed slices `&[TypeName]` for arrays.

Field on the other hand are owned values. When  rendering
fields we will try to avoid boxing. For now structs with
simple field types are owned.

Also added timing to the different phases of generation.