Add type_name_ utility to DataObject
It is necessary sometimes to get the type name as String from DataObjects.
For example to process events one need the name of the data type.
To this effect a new synthetic method is added to the DataObject.

The tests and eventster example have been revised to use the new method.