Add sorting of data types
Use pre-order traversal of the hierarchy as to sort elements suc that each parent and all of its children are in contiguous block. This allows dispatching efficiently s11n tasks based on match statements that can be optimized ot jump maps. The same can be used to determine if one type is child of another. The same can be used to dispatch cross trait type conversions.