Purge MethodFault and Event hierarchies
This change introduces ability in the vim model loader and generator to skip the generation of type sub-hierarchies.

A list of root types whose hierarchies are purged is passed to the vim model loader. The loader than marks the data types that will be skipped and those whose hierarchies are purged.

The various emitters are thought how to deal with this to skip unnecessary types and enhance the pruned types as to provide dynamic access to the wire discriminator and   extra fields.