Allow types checks on skipped types
MethodFault and Event have in places deep and broad hierarchies. It thus may be very hard
to check for if an error is a base class of a big subtree of classes. To simplify this the
type_name_ is change to type_ that is a StructType enum not a string anymore. StructType is
extended to have all known types not only the generated types. Further `child_of()` method
is added to `StructType` allowing one to check if a given type belongs to a parent
hierarchy e.g. is a given fault descendant of `VmConfigFault`. For example

```rust
fault.type_.child_of(StructType::VmConfigFault)
```