Basic Task support in vTUI
The current implementation shows recent tasks as resource type and also for individual objects by pressing "t".

There are no hints. As we need to figure out the hints location for tasks, alarms, events and permissions.

This change adds Tasks to vim_macros

A bug is found that selecting MethodFault properties with vim_macros fails. It seems the type is rendered as a Boxed trait not Structure.