Fix fields of wrong type
Several fields in the OpenAPI spec are declared as struct field while in the XSD schemas those are declared as String. This was causing serde_json unmarshalling errors.

This change patches up the OpenAPI spec for several such instances found to be different in the Broadcom API docs between the VI/JSON and the SOAP API docs.