Add vSphere 9.0 API fixes to the generator and updated bindings
This changes fixes the generator to support new style reference type properties.
In the new style properties with description have "allOf" with one element that
is a reference. The actual schema contains only the `description` and `allOf`.
In the older schema the description was nested in the reference and this was
violating the OpenAPI 3.0 format. As of this change both old and new formats are
supported.

New bindings have been generated. Those include vsan, pbm, vslm, sms and eam in
addition to VIM. The Client is not fixed to support these additional modules.
Specifically the references to the well known root objects like ServiceInstance
are not added to the Client object. Those values can perhaps be scavenged from
the VCF Java SDK, govmomi or pyvmomi.

New versions were manually added to the client 8.0.3.0 and 9.0.0.0 as published
in the respective VI JSON specs.

A fix was needed to the event example to accommodate new field.

Removed old specs. Checked that equivalent of
https://github.com/noclue/vim_rs/commit/d9e4a712021922f724139a450de6cc77a270f319
is present in the updated 9.0.0.0 spec.