Remove the DataObject trait methods
This is minor optimization to remoce the data object trait methods as those do not bring value. Also the DataObject field is removed from child object as well as any dataobject descendants that have no fields of their own. The last changes coding much simpler as seen in the examples.

Unfortunately this does not seem to improve build times at all.