Introduce TaskTracker utility
This change introduces a TaskTracker utility that can be used to monitor for task completion using Rust native primitives.

As TaskTracker works in the background the property collector cache object is now converted to thread friendly and using `Arc<RwLock<_>>` instead of `Rc<RefCell<_>>`. This also invalidates the `SharedRefProxy` utility.

The `vm_rename` example shows how to use use the TaskTracker utility.